Tp5 done

+++ app/Http/Controllers/ProductController.php
@@ -2,57 +2,65 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Product;
 use Illuminate\Http\Request;
+use App\Models\Product;
 
 class ProductController extends Controller
 {
     public function getProducts()
     {
         $products = Product::all();
-        return response()->json([
-            "message" => "Getting all products",
-            "data" => $products
-        ]);
+        return response()->json($products);
     }
 
     public function createProduct(Request $request)
     {
         $validated = $request->validate([
             'name' => 'required|string|max:255',
+            'category_id' => 'required|integer|exists:categories,id',
+            'price' => 'required|integer',
+            'description' => 'sometimes|string',
+            'images' => 'sometimes|json ',
         ]);
-
         $product = Product::create($validated);
-
-        return response()->json([
-            "message" => "Creating 1 new product",
-            "data" => $product
-        ], 201);
+        return response()->json($product, 201);
     }
 
+    
+    public function getProduct($productId)
+    {
+        $product = Product::find($productId);
+        if (!$product) {
+            return response()->json(['error' => 'Product not found'], 404);
+        }
+        return response()->json($product);
+    }
+
+    
     public function updateProduct(Request $request, $productId)
     {
-        $product = Product::findOrFail($productId);
-
+        $product = Product::find($productId);
+        if (!$product) {
+            return response()->json(['error' => 'Product not found'], 404);
+        }
         $validated = $request->validate([
-            'name' => 'sometimes|required|string|max:255',
+            'name' => 'required|string|max:255',
+            'price' => 'required|integer',
+            'category_id' => 'required|integer|exists:categories,id',
+            'description' => 'sometimes|string',
+            'images' => 'sometimes|json',
         ]);
-
         $product->update($validated);
-
-        return response()->json([
-            "message" => "Updating product with id: $productId",
-            "data" => $product
-        ]);
+        return response()->json($product);
     }
 
     public function deleteProduct($productId)
     {
-        $product = Product::findOrFail($productId);
+        $product = Product::find($productId);
+        if (!$product) {
+            return response()->json(['error' => 'Product not found'], 404);
+        }
         $product->delete();
-
-        return response()->json([
-            "message" => "Deleting product with id: $productId"
-        ]);
+        return response()->json(['message' => 'Product deleted']);
     }
 }
